Move generic debugger client code to mlc
