+++ packages/examples/resources/debugger/Dockerfile
@@ -2,8 +2,8 @@ FROM ghcr.io/graalvm/graalpy-community:24
 
 RUN curl https://get.volta.sh | bash
 ENV VOLTA_FEATURE_PNPM=1
-ENV VOLTA_HOME "/root/.volta"
-ENV PATH "$VOLTA_HOME/bin:$PATH"
+ENV VOLTA_HOME="/root/.volta"
+ENV PATH="$VOLTA_HOME/bin:$PATH"
 
 RUN volta install node@22
